requirements.txt
gunicorn==19.9.0
astroid==2.3.3
astropy==3.2.3
astunparse==1.6.3
atomicwrites==1.3.0
attrs==19.3.0
bokeh==1.4.0
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
distributed==2.8.0
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
Flask==1.1.2
heroku==0.1.4
Jinja2==2.10.3
numpy==1.17.3
pandas==0.25.3
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.9.0
selenium==3.141.0
spacy==2.3.2
sparse==0.8.0
joblib==0.14.0
Sphinx==2.2.1
urllib3==1.24.2
python-Levenshtein==0.12.0
simplejson==3.17.2
json5==0.8.5
matplotlib==3.1.1
